refactor(render): Clarify redunant mod is for graphical renderer
